requirements.txt
# Core Agent Framework
agent-framework>=0.1.0

# HTTP Requests
requests>=2.31.0

# Environment Configuration
python-dotenv>=1.0.0

# MCP Server Support
fastmcp>=0.1.0

# Data Validation
pydantic>=2.0.0

# OpenTelemetry - Pin versions to avoid conflicts
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-exporter-otlp==1.39.1

# Optional: Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Optional: Excel Export
openpyxl>=3.1.0

# Optional: Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Optional: Advanced Analytics
scipy>=1.10.0
scikit-learn>=1.3.0

# Optional: Database Support
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# redis>=4.5.0

# Optional: Async HTTP (for faster API calls)
# aiohttp>=3.8.0
# httpx>=0.24.0

# Optional: Web Scraping
# beautifulsoup4>=4.12.0
# lxml>=4.9.0
# selenium>=4.10.0

# Optional: Testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0

# Development Tools
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.4.0
